Issue #11163: Enforcing file size on AbstractJavadocPosition
