+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/abstractjavadoc/InputAbstractJavadocPositionOne.java
@@ -1,0 +1,108 @@
+/*
+com.puppycrawl.tools.checkstyle.checks.javadoc.AbstractJavadocCheckTest$JavadocCatchCheck
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.javadoc.abstractjavadoc;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/////////////
+// CLASSES //
+/////////////
+
+/**Javadoc*/
+class/**nope*/ InputAbstractJavadocPositionOne/**nope*/{
+    /**Javadoc*/
+    protected/**nope*/ class/**nope*/ B/**nope*/{/**nope*/}
+
+    /**Javadoc*/
+    private/**nope*/ static/**nope*/ class/**nope*/ C/**nope*/{/**nope*/}
+
+    /**Javadoc*/
+    @Component/**nope*/ class/**nope*/ D/**nope*/{/**nope*/}
+
+    /**Javadoc*/
+    @Component/**nope*/ private/**nope*/ class/**nope*/ E/**nope*/{/**nope*/}
+
+    /**Javadoc*/
+    private/**nope*/ @Component/**nope*/ class/**nope*/ F/**nope*/{/**nope*/}
+}
+
+//////////////////
+// CONSTRUCTORS //
+//////////////////
+
+/**Javadoc*/
+class/**nope*/ AA/**nope*/{
+    /**Javadoc*/
+    AA/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+/**Javadoc*/
+class/**nope*/ BB/**nope*/{
+    /**Javadoc*/
+    private/**nope1*/ BB/**nope2*/(/**nope3*/)/**nope4*/{/**nope5*/}/**nope6*/
+}
+
+class/**nope*/ DD/**nope*/{
+    /**Javadoc*/
+    @Component/**nope*/ DD/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class/**nope*/ EE/**nope*/{
+    /**Javadoc*/
+    @Component/**nope*/ private/**nope*/ EE/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class/**nope*/ FF/**nope*/{
+    /**Javadoc*/
+    private/**nope*/ @Component/**nope*/ FF/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+/////////////
+// METHODS //
+/////////////
+
+class AAA {
+    /**Javadoc*/
+    void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class BBB {
+    /**Javadoc*/
+    private/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class CCC {
+    /**Javadoc*/
+    static/**nope*/ private/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class DDD {
+    /**Javadoc*/
+    @Component/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class EEE {
+    /**Javadoc*/
+    @Component/**nope*/ private/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class FFF {
+    /**Javadoc*/
+    static/**nope*/ @Component/**nope*/ private/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class GGG {
+    /**Javadoc*/
+    void/**nope*/ a/**nope*/(@Component/**nope*/int/**nope*/ a/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class HHH {
+    /**Javadoc*/
+    java.lang.String/**nope*/ a/**nope*/()/**nope*/{/**nope*/return null;/**nope*/}/**nope*/
+}
